Update debug paths, move code to common.utils
